test(lineStation): add unregistred station test
