+++ src/test/java/nextstep/subway/line/acceptance/LineStationRemoveAcceptanceTest.java
@@ -14,6 +14,7 @@ import java.time.LocalTime;
 
 import static nextstep.subway.line.acceptance.step.LineAcceptanceStep.지하철_노선_등록되어_있음;
 import static nextstep.subway.line.acceptance.step.LineStationAddAcceptanceStep.*;
+import static nextstep.subway.line.acceptance.step.LineStationRemoveAcceptanceStep.*;
 import static nextstep.subway.station.acceptance.step.StationAcceptanceStep.지하철역_등록되어_있음;
 
 @DisplayName("지하철 노선에서 역 제외 기능")
@@ -80,5 +81,16 @@ public class LineStationRemoveAcceptanceTest extends AcceptanceTest {
         // then
         지하철_노선에_지하철역_제외_확인됨(response, stationId2);
         등록된_지하철역이_정렬되어_위치됨(response, Lists.newArrayList(1L, 3L));
+    }
+
+    @DisplayName("지하철 노선에서 등록되지 않는 역을 제외한다.")
+    @Test
+    void removeStationNotCreate() {
+        Long notCreatedStationId = 4L;
+        // when
+        ExtractableResponse<Response> deletedStationResponse = 지하철_노선에_지하철역_제거_요청(lineId, notCreatedStationId);
+
+        // then
+        지하철_노선에_지하철역_제외_실패됨(deletedStationResponse);
     }
 }
